The Stylesheet update,
Add - New stylesheet called normalize to replace the asterisk I was
previously using. I figured it is better practice to use this method.
Change - Formatted the main.css to make it more readable
Change - Index.html to include the link rel to the normalize file
Del - database structures document, due to it not being relevant